Add files via upload
Normal distribution.  Code to find the area under a curve in a given range, also: mean , deviation and an untested method to find the x-bar distribution.

The function using SciPy is the function to use.
function mycdf() is not working accurately and am struggling to find where the numbers are going a bit wonky compared to the SciPy numbers.